+++ .github/workflows/MainDistributionPipeline.yml
@@ -18,7 +18,7 @@ jobs:
     with:
       duckdb_version: v1.5.1
       ci_tools_version: v1.5.1
-      extension_name: quack
+      extension_name: waddle
 
   code-quality-check:
     name: Code Quality Check
@@ -26,5 +26,5 @@ jobs:
     with:
       duckdb_version: v1.5.1
       ci_tools_version: v1.5.1
-      extension_name: quack
+      extension_name: waddle
       format_checks: 'format;tidy'
